fix: parquet2 build
