+++ rust/src/kernel/error.rs
@@ -6,6 +6,7 @@ pub type DeltaResult<T, E = Error> = std::result::Result<T, E>;
 #[derive(thiserror::Error, Debug)]
 #[allow(missing_docs)]
 pub enum Error {
+    #[cfg(feature = "arrow")]
     #[error("Arrow error: {0}")]
     Arrow(#[from] arrow_schema::ArrowError),
